requirements.txt
aiohttp>=3.9.0
aiodns>=3.1.0
pyyaml>=6.0.1
python-telegram-bot>=20.7
apscheduler>=3.10.4
python-dateutil>=2.9.0.post0
python-dotenv>=1.0.1
feedparser>=6.0.11
regex>=2024.11.6
tenacity>=8.3.0
pydantic>=2.6.4
